+++ DoNetDrive.Protocol.Elevator/Door8864/Card/WriteCardList/07CardListBySort/WriteCardListBySort.cs
@@ -32,7 +32,7 @@ namespace DoNetDrive.Protocol.Elevator.FC8864.Card.CardListBySort
         /// </summary>
         /// <param name="CardList">错误卡列表</param>
         /// <param name="buf"></param>
-        protected override void ReadCardByFailBuf(List<ulong> CardList, IByteBuffer buf)
+        protected override void ReadCardByFailBuf(List<decimal> CardList, IByteBuffer buf)
         {
             CardList.Add((UInt64)buf.ReadInt());
         }
